bot-handler/requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.1.0
certifi==2024.7.4
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
pip==22.3.1
pyaes==1.6.1
pyasn1==0.6.0
pydantic-core==2.20.1
pydantic==2.8.2
python-dotenv==1.0.1
rsa==4.9
setuptools==65.5.1
SQLAlchemy==2.0.31
# Telethon==1.36.0
typing-extensions==4.12.2
wheel==0.38.4
yarl==1.9.4
